more issues with sample reordering and simplifying compilation
